Create BLorient11601 (#2259)
* Create BLorient11601

I created a record for Or. 11601 according to Strelcyns catalogue. I did not have any images for comparison, but I think, it is safe to assume two codicological units, given the fact that dimensions of folia and hands are changing after folio 76. For the numerous additions, I do not know, whether they have been added before or after the latest binding. Because this is not clear, I found it more safe to add them to the physdesc of the Codex, but not to one of the codicological units.

* Updated BLorient11601

I added some remarks, that escaped my attention in my first pull request a couple of minutes before.

* Update BLorient11601

I updated the record according to Eugenias and Denis requests.

* Update BLorient11601

I updated the record according to Eugenias requests.

* Update BLorient11601

I updated the record according to Eugenias request.