Update BLorient8192
I updated BLorient8192. I decided to rework it as a composite manuscript to document the details given for p2 in a searchable and decent way.

I did not find a LIT ID for a salām to Mary beginning with ሰላም፡ ለማርያም፡ ታቦት፡ Probably, I can create a new one.
There is no INS ID for the British Red Cross Society, who benefited from the purchase of the manuscript to the British Museum. Do you think, it is appropriate to create a INS or a LOC ID for the British Red Cross Society?