Create BLorient8819
I created a record for BLorient8819 according to the description of St. Strelcyn.

Strelcyn subdivided the main text different to what was encoded by Denis to LIT6193HomiliaryPassions. The subdivision in LIT6193HomiliaryPassions is based on the witness BNFet116. However, either in BNFet116 nor in the incipits given by Strelcyn for BLorient8819, I found evidence to verify or falsify this subdivision. Probably the subdivision was based on the printed editions. In the end, I do not know, whether the two manuscripts are different in their attribution of textparts to the days of the week. Yet alone the existence of ms_i1.3 seems to be unique to BLorient8819. I suggest to create a new sub-ID or LIT-ID for it.

For ms_i2, see our discussion in https://github.com/BetaMasaheft/Documentation/issues/2562#issuecomment-2174464848.