Create BLorient11719
I created a record according to the catalogue of St. Strelcyn. I was confused about the composition of the manuscript. Strelcyn has record different dimensions, layouts, hands and contents. However, changes in dimension, layout, hand or content do not always correlate with each other. It therefore under question, whether it is a composite manuscript overall and if yes, which parts can be discerned.

I was not able to identify the texts in p2 and opened two issues: https://github.com/BetaMasaheft/Documentation/issues/2565 and https://github.com/BetaMasaheft/Documentation/issues/2566