Create BLorient13273 according to Strelcyn (#2340)
* Create BLorient13273 according to Strelcyn

I created a record according to S. Strelcyns catalogue. However, I was not able to identify ms_i6 (https://github.com/BetaMasaheft/Documentation/issues/2508), ms_i8 (https://github.com/BetaMasaheft/Documentation/issues/2509) and a2. For the first and the second, I suggest to create new LIT IDs as argued in the issues. The magical prayer in a2 is probably also a candidate for a new LIT ID, but I did not find any other witness. I checked manuscripts containing LIT4302Prayer and LIT5876AsmatPrayer, but to my point of view, this piece is to much different to identify it as one of these.

* Update BLorient13273

I forgot to push the latest update of ms_i3.

* Update BLorient13273 (semikolons) according to Dorotheas request.

* Update BLorient13273 according to our discussion

... in https://github.com/BetaMasaheft/Documentation/issues/2508.

* Update records

I updated the record BLorient13273 with the new LIT ID for the Prayer on the Passion of Jesus Christ as discussed in https://github.com/BetaMasaheft/Documentation/issues/2509#issuecomment-2065221614. In addition I updated also two other witnesses.

* Update EMIP01909 (add semicolons in title)

Minor corrections according to Dorotheas requests (added semicolons in title)