Create BLorient13267 (#2437)
* Create BLorient13267

There is no sub-ID for an introduction in the record of LIT2200Qalaha. However Strelcyn recorded a fairly extensive introduction from 1ra to 3rb (alas without incipit). In both witnesses linked to LIT2200Qalaha, there are also introductions of similar size attested as in BDLaethf9 and InEtn210317a . I propose therefore to update LIT2200Qalaha and add a sub-ID "introduction".

The problem of the so-called "unbound" is similar to https://github.com/BetaMasaheft/Manuscripts/pull/2431/files. For this reason, I made the same comment in the binding description.

* Update BLorient13267 according to Eugenias requests

* Update BLorient13267 (updated main language and deleted additions)