Create BLorient8818 (#2436)
I created a record for BLorient8818. This is a manuscript, in which a text of Śǝna fǝṭrat (LIT2314senafe) is placed in between the first and the second chapter of LIT1097Ammest. I was puzzled, whether this text should be nested inside ms_i1 as part of LIT1097Ammest or kept independantly.
For ms_i4, I do not know, whether it is reasonable to create a general record for "List of Patriarchs of Alexandria". Since there is no further information, I do not know, whether it is probably related to the accounts of Walda Amid or Abušāker.
I found information on B. (= J.) R. T. Lieder. I understood this, as Lieder bought this manuscript and sold it to Curzon, who brought it to England.