Create BLorient13275 (#2434)
* Create BLorient13275

I created a record for BLorient13275. I could not identify the text, that is entitled "The hymn-invocation Lǝsāna sabʾ" by Strelcyn as discussed in https://github.com/BetaMasaheft/Documentation/issues/2547.

For the cylindrical container, I encoded it as wood, because, we do not have a keyword for bamboo.

I came across the record EMIPms00473 and I found, that width: 7.0 could be read as width 7,0. I think, this must be 70 mm, because otherwise it can hardly carry any text.

* Update BLorient13275 (added incomplete to title in msItem)

* Update BLorient13275

according to our discussion in https://github.com/BetaMasaheft/Documentation/issues/2547.