Create BLorient13278 (#2385)
* Create BLorient13278

I created a record according to S. Strelcyns description. I was able to identify the msItems, except for ms_i2, for which I propose a new LIT ID, as I have outlined in https://github.com/BetaMasaheft/Documentation/issues/2520.

Unfortunatly, I made a mistake in creating a record with the name BL13278 instead of BLorient13278 as it is supposed to be. I created another new record with the correct name. The outdated earlier record can be ignored in this pull request, but should be deleted in a new pull request immediately after merging this one. When I remember it correctly, a new pull request is needed, because otherwise, the deletion will cause turmoil in the app. I this this correct?

* Update BLorient13278 according to Dorotheas requests

* Update BLorient13278.xml

* Update BLorient13278 and delete BL13278

I updated the record BLorient13278 and deleted BL13278 as requested by Eugenia and Dorothea.

---------

Co-authored-by: thea-m <thea.m@hotmail.de>