Update BLorient11719
I obstained from creating a new LIT ID for the text attested in p2_i2, because I do not know, whether this text of an explanation in Amharic has an independent circulation and I think, it does not need an own ID. However, I would quickly create one, if you find that is neccessary.

I updated the markup in p2_i3 according to what I noted in https://github.com/BetaMasaheft/Documentation/issues/2566.