Create BLorient11719 (#2491)
* Create BLorient11719

I created a record according to the catalogue of St. Strelcyn. I was confused about the composition of the manuscript. Strelcyn has record different dimensions, layouts, hands and contents. However, changes in dimension, layout, hand or content do not always correlate with each other. It therefore under question, whether it is a composite manuscript overall and if yes, which parts can be discerned.

I was not able to identify the texts in p2 and opened two issues: https://github.com/BetaMasaheft/Documentation/issues/2565 and https://github.com/BetaMasaheft/Documentation/issues/2566

* Update BLorient11719 according to Jonas requests

* Update BLorient11719 according to Jonas request

* Update BLorient11719

I updated the record and added new LIT for p2_i1 as discussed in https://github.com/BetaMasaheft/Documentation/issues/2565.

Other issue as new LIT for p2_i2 and p2_i3 are still open, the latter to be discussed in https://github.com/BetaMasaheft/Documentation/issues/2566.

* Update BLorient11719

I obstained from creating a new LIT ID for the text attested in p2_i2, because I do not know, whether this text of an explanation in Amharic has an independent circulation and I think, it does not need an own ID. However, I would quickly create one, if you find that is neccessary.

I updated the markup in p2_i3 according to what I noted in https://github.com/BetaMasaheft/Documentation/issues/2566.