Create BLorient5085
Some leaves have been moved. This seems to have led to some confusion on the part of St Strelcyn, who wrote several chapters as well as the index on folio 2ra and following. This must be a mistake, because folios 2 to 11 cannot carry both the index and chapters 10 to 14. However, I have decided to mark the uncertainty and present a possible solution. It can be checked and corrected when images of the BL are available, hopefully in September.