Create BLorient5426
I was not able to mark up the prayers in LIT1758Lefafa#MangadaSamay. The order does not correspond to the ordering in Euringer and the incipits are too short and have too much variation to stick them to the existing sub-ID (i.e. LIT1758Lefafa#FirstPrayer, etc.).
The last msItem (ms_i3) was described as "Developed formula of benediction given by God to the Archangel Michael for the scribe, the owner, etc. of this manuscript." I think, he did not realize, that it is also some part of LIT1758Lefafa, as can be discerned by comparison with the transcribed text of BMLacq384 (on f. 9).