Create BLorient13274
Again, I did not markup the prayers of Mangada samāy individually. Given the number 7 and the title (ጸሎት፡ ዘመንገደ፡ ሰማይ፡), it is very likely, that these are the same as edited in Euringer, 1940. However, the experience with BLorient5426 was teaching me, that the order might be different, what I cannot verify from the catalogue description.