Update BLorient7946.xml (#2568)
* Update BLorient7946.xml

According to Denis request in https://github.com/BetaMasaheft/Manuscripts/pull/2516#pullrequestreview-2268162385.

* Update BLorient7946.xml

I added the information on folio-number according to Denis request.
In addition I revised the record. I added ms_i1.1 that I did notice before and I defined a loosely inserted piece of paper (as I understand) as accMat. I hope, that the definition of this accMat as xml:id="accMat1" (to refer to it as an anchor) was correct. I did not find explicit information on xml:id for accMat in the Guidelines. However, so far the oxygen did not protest against it (neither to any other anchor, that I inserted for test reasons). I am not sure, if the possibility to define accMat with xml:id has been discussed before.

* Update BLorient7946 according to Denis request

* Update BLorient7946 according to Denis request