Create BLorient13274 (#2521)
* Create BLorient13274

Again, I did not markup the prayers of Mangada samāy individually. Given the number 7 and the title (ጸሎት፡ ዘመንገደ፡ ሰማይ፡), it is very likely, that these are the same as edited in Euringer, 1940. However, the experience with BLorient5426 was teaching me, that the order might be different, what I cannot verify from the catalogue description.

* Update BLorient13274

According to Denis remarks in https://github.com/BetaMasaheft/Manuscripts/pull/2520#pullrequestreview-2215051372.

* Update BLorient13274 according to Jonas remark