Create BLorient12996 (Amharic grammar) (#2505)
* Create BLorient12996 (Amharic grammar)

I encoded BLorient12996 according to the description of St. Strelcyn. However, Strelcyn was not interested in the first part, containing an Arabic grammar, but gave detailed information only on the second part with an Amharic grammar. I would like to create a new LIT ID for this latter text, as I argued in
https://github.com/BetaMasaheft/Documentation/issues/2575. Otherwise, I would remove the placeholder IDs (LIT0000000000000#Script, etc.) before merging.

* Update BLorient12996 according to Jonas and Eugenias requests

Including update of STUD0009AmharicGram.

* Update BLorient12996 (inserting figure)

As discussed in https://github.com/BetaMasaheft/Documentation/issues/2569#issuecomment-2232655971 and described in https://betamasaheft.eu/Guidelines/?q=figure&id=FiguresImages.

* Update BLorient12996 according to Denis' requests