Create BLorient11804
To my opinion the incipits of ms_i1 and ms_i2 do not contain enough text to make it possible to these prayers with any other record except for the generic records LIT5696MagicPrayer and LIT5888AsmatPrayer respectively. Do you agree, or do you think, it is warranting to create issues for them and look for parallels and eventually create new IDs?