Update BLorient8818
I decided to delete the placeholder tag. I did not find any suitable LIT or NAR ID. To establish a dependancy or relation to possible sources or manuscript texts, it would require access to the images of BLorient8818 and some time for research. Without checking such relations and without any incipit, I think, it makes no sense to create a new CAe.