Add change notes to mss records_no.7 (#2970)
The seventh and last lore of manuscript, for which I added a change element, so they will not be counted as drafts or stubs any more. For the remaining handful of records, the label "stub" or "draft" is legitimate, because, they contain just a very few information. cf. https://github.com/BetaMasaheft/Documentation/issues/2976.