Draft BLorient12034_CH (#2980)
* Draft BLorient12034_CH

I created a draft for BLorient12034 with a setup of codicoligical parts and textparts as described in the catalogue of St. Strelcyn. I still need to investigate many LIT ID, that are necessary for a full description. In order not to have too many issues open in the same time and keep this record reviewable, I want to to merge it with this basic information and enhance the record later.

* delete otherLang="am" from p2_i75

* Update according to Magdas remarks