Create BNUet5.xml
I created a record for BNUet5. Contrary to my announcements in https://github.com/BetaMasaheft/Documentation/issues/3054, I did not suffice myself with creating a stub, but added  information available to me from the listed catalogues, my own PhD thesis and from images online. I found out, that the Tarika Walda Amid-text is not part of the α-recension, but probably an interpolated version of α and β with the basis of the β-recension.

The PRS record for the scribe Habta Mikael is still missing. I create a pull request in person as soon as this will be possible again hopefully in the next week.