Added internal dates, additions
I updated the note on coloph1 and calculated and marked up the dates and added evidence to origin and handnote. I committed myself once again and added additions to the record to understand better the statements by Zanutto and Ewald. I hope, that it is more correct and clear, now.