Create BNUet5.xml (#3041)
* Create BNUet5.xml

I created a record for BNUet5. Contrary to my announcements in https://github.com/BetaMasaheft/Documentation/issues/3054, I did not suffice myself with creating a stub, but added  information available to me from the listed catalogues, my own PhD thesis and from images online. I found out, that the Tarika Walda Amid-text is not part of the α-recension, but probably an interpolated version of α and β with the basis of the β-recension.

The PRS record for the scribe Habta Mikael is still missing. I create a pull request in person as soon as this will be possible again hopefully in the next week.

* Corrects after Dorotheas review

* update foliation

* Up according to Nafisas requests

* Up according to Nafisas requests

* Added internal dates, additions

I updated the note on coloph1 and calculated and marked up the dates and added evidence to origin and handnote. I committed myself once again and added additions to the record to understand better the statements by Zanutto and Ewald. I hope, that it is more correct and clear, now.

* Update LOC1464Ankoba (town)