Created BNFet391 and BNFet654
I created two draft versions for  BNFet391 and BNFet654, i.e. I described the content of the two manuscripts as it is presented by Strelcyn in his catalogue of 1954, but did not consult any images and did not look up CAe or PRS IDs.
The records are necessary to refer to it in some newly created work record and I think, there will be more than one more case in future.
I made one correction and two updates to BNFet390.