Update mss with divinatory texts (#3088)
* Update mss with divininatory texts

I updated a number of manuscript records with divinatory texts (ḥassābs), which are referenced in Strelcyns description on BLorient12034. I created a number of records for ḥassābs, which I wanted to include also in the records of CamOr1885, BNFet390, BNFet391, BNFet405, IVorlov15 and BAVet244, all of which I have updated without having a look into the images and some of which the physDesc and msContents are not complete, yet.

* Correct error CamOr1885

* Finished BNFet405 description

Corrected error which prevented validation and completed msContents, PhysDesc and history.

* Update change note BNFet405

* Add lang Latin to BAVet244

* Updates according to Magdas suggestions and requests