Create ANLcr90 (#3140)
* Create ANLcr90.xml

I believe, that it would be an improvement to update LIT1419GadlaA and LIT5061TaammeraA and create a LOC ID for Zāʾtā in ʾƎndartā, as I have indicated it in the record.

* LOC6450Zatta in comment

* Updates according to Magdas and Eugenias requests

* Add sub IDs to ms_i1.2

I updated ANLcr90 with sub IDs for ms_i1.2 (LIT5061TaammeraA) after checking bm:Lusini1990Ananya (cf. https://github.com/BetaMasaheft/Works/pull/1854). The edition by Lusini (1990) was based on this manuscript as codex unicum. In the meantime a second witness (ESky066) has become known.

* Validate record

* Shorten note on Gadla Ananya, add bibl Lusini

* Add history

* Update origin