Update ANLcr76.xml (#3149)
* Update ANLcr76.xml

I created a record for ANLcr76.

@DenisNosnitsin1970 may have a look on the tigrinya-texts and transcription, if time and energy allow.

I had difficulties to determine the number of folios in both textparts, because Strelcyns description switches from pages to folios and mentions endleaves en passant and does not make clear statements on the number of endleaves. I hope, that my description is as clear as possible and necessary.

* Updates according to Magdas requests